Sample/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
murf==2.0.0
requests==2.31.0
assemblyai==0.43.1
google-generativeai==0.3.2
pymongo==4.6.0
motor==3.3.2
